feature: Add php/wasmedge-php-7.4.32 Makefile target
This target generates a specific PHP CLI targeting wasmedge. This
allows for this WebAssembly module to, e.g. perform a `socket` call in
order to create a listening socket as opposed to a socket being
created by the host, and the file descriptor being forwarded to the
WebAssembly module.

It also exports  a `connect` function that resembles `connect(3)`.

The `WASMLABS_RUNTIME` is exposed to the build scripts in order to
produce the specified target.