feature: allow to skip wasm-opt when invoked by clang
By exporting the WASMLABS_SKIP_WASM_OPT envvar, the wasm-opt wrapper
present in the wasm-base container image will make the wasm-opt call a
no-op.

This is useful for cases when we don't desire wasm-opt to be called,
for example when invoked directly by
clang (https://github.com/llvm/llvm-project/issues/55781).